--- src/com/acbelter/Utils.java
@@ -1,30 +1,0 @@
-package com.acbelter;
-
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
-
-public final class Utils {
-    private Utils() {
-    }
-
-    public static String generatePasswordHash(String password) {
-        MessageDigest md;
-        try {
-            md = MessageDigest.getInstance("SHA-256");
-        } catch (NoSuchAlgorithmException e) {
-            return null;
-        }
-        md.update(password.getBytes());
-
-        byte[] byteData = md.digest();
-        StringBuilder sb = new StringBuilder();
-        for (byte b : byteData) {
-            String hex = Integer.toHexString(0xff & b);
-            if (hex.length() == 1) {
-                sb.append('0');
-            }
-            sb.append(hex);
-        }
-        return sb.toString();
-    }
-}
